cleaned and organized My library
